+++ tests/fixtures/dependency/extra/class-example-class.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace ComposerWordPressAutoloaderTests_Dependency\Extra;
+
+class Example_Class
+{
+}