+++ tests/fixtures/dependency/src/class-example-class.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace ComposerWordPressAutoloaderTests_Dependency;
+
+class Example_Class
+{
+}