+++ tests/fixtures/dependency/tests-extra/class-example-test-file.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ComposerWordPressAutoloaderTests_Dependency\Extra\Tests;
+
+class Example_Test_File
+{
+
+}