+++ tests/fixtures/root/tests/class-example-test-file.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ComposerWordPressAutoloaderTests\Tests;
+
+class Example_Test_File
+{
+
+}