+++ CHANGELOG.md
@@ -5,6 +5,11 @@ documented in this file.
 
 ## [Unreleased]
 
+## 0.2.0
+
+- Updates autoloader to use non-hard-coded paths.
+- Adds support for dependencies to autoload files as well, fixes [#3](https://github.com/alleyinteractive/composer-wordpress-autoloader/issues/3).
+
 ## 0.1.0
 
 - Initial release.